src/ml/requirements.txt
flask==2.3.3
flask-cors==4.0.0
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
